+++ codegen/src/main/java/software/amazon/awssdk/codegen/poet/rules/TestGeneratorUtils.java
@@ -27,8 +27,8 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import software.amazon.awssdk.awscore.rules.AwsEndpointAttribute;
-import software.amazon.awssdk.awscore.rules.SigV4AuthScheme;
-import software.amazon.awssdk.awscore.rules.SigV4aAuthScheme;
+import software.amazon.awssdk.awscore.rules.authscheme.SigV4AuthScheme;
+import software.amazon.awssdk.awscore.rules.authscheme.SigV4aAuthScheme;
 import software.amazon.awssdk.codegen.model.rules.endpoints.ExpectModel;
 import software.amazon.awssdk.core.rules.model.Endpoint;
 import software.amazon.awssdk.core.rules.testing.model.Expect;
