+++ core/aws-core/src/main/java/software/amazon/awssdk/awscore/rules/authscheme/EndpointAuthScheme.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.awscore.rules;
+package software.amazon.awssdk.awscore.rules.authscheme;
 
 import software.amazon.awssdk.annotations.SdkProtectedApi;
