+++ test/codegen-generated-classes-test/src/test/java/software/amazon/awssdk/services/endpointproviders/EndpointAuthSchemeInterceptorTest.java
@@ -23,11 +23,9 @@ import static org.mockito.Mockito.mock;
 import java.util.Arrays;
 import org.junit.Test;
 import software.amazon.awssdk.auth.signer.Aws4Signer;
-import software.amazon.awssdk.auth.signer.AwsSignerExecutionAttribute;
-import software.amazon.awssdk.awscore.AwsExecutionAttribute;
 import software.amazon.awssdk.awscore.rules.AwsEndpointAttribute;
-import software.amazon.awssdk.awscore.rules.SigV4AuthScheme;
-import software.amazon.awssdk.awscore.rules.SigV4aAuthScheme;
+import software.amazon.awssdk.awscore.rules.authscheme.SigV4AuthScheme;
+import software.amazon.awssdk.awscore.rules.authscheme.SigV4aAuthScheme;
 import software.amazon.awssdk.core.SdkRequest;
 import software.amazon.awssdk.core.interceptor.ExecutionAttributes;
 import software.amazon.awssdk.core.interceptor.ExecutionInterceptor;
